move the proofgen and txbuilder mocks added in this PR into the central internal/tests/mocks package (Attestor's mock stays local -- centralizing it creates an import cycle through internal/chains)
